+++ package.json
@@ -7,6 +7,7 @@
     "@material-ui/icons": "^4.9.1",
     "@material-ui/styles": "^4.9.0",
     "clipboard-copy": "^3.1.0",
+    "esm": "^3.2.25",
     "gatsby": "^2.19.17",
     "gatsby-plugin-google-analytics": "^2.1.35",
     "gatsby-plugin-manifest": "^2.2.41",
