+++ src/components/AbilitiesColumnProvider.js
@@ -7,6 +7,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function AbilitiesColumnProvider({ children, currentBook }) {
   let bookData = ProvideBookData()
 
@@ -16,7 +18,11 @@ export default function AbilitiesColumnProvider({ children, currentBook }) {
       name: "name",
       options: {
         customBodyRender: (value, tableMeta) => (
-          <Link to={`/abilities/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}>
+          <Link
+            to={`/abilities/${
+              tableMeta.rowData[GENERATED_ID_COL_INDEX]
+            }/${slugify(value)}/`}
+          >
             {value}
           </Link>
         ),
