+++ src/components/AdversariesArmorColumnProvider.js
@@ -8,6 +8,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function AdversariesArmorColumnProvider({
   children,
   currentBook,
@@ -21,7 +23,9 @@ export default function AdversariesArmorColumnProvider({
       options: {
         customBodyRender: (value, tableMeta) => (
           <Link
-            to={`/adversaries-armor/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}
+            to={`/adversaries-armor/${
+              tableMeta.rowData[GENERATED_ID_COL_INDEX]
+            }/${slugify(value)}/`}
           >
             {value}
           </Link>
