+++ src/components/AdversariesGearColumnProvider.js
@@ -8,6 +8,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function AdversariesGearColumnProvider({
   children,
   currentBook,
@@ -21,7 +23,9 @@ export default function AdversariesGearColumnProvider({
       options: {
         customBodyRender: (value, tableMeta) => (
           <Link
-            to={`/adversaries-gear/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}
+            to={`/adversaries-gear/${
+              tableMeta.rowData[GENERATED_ID_COL_INDEX]
+            }/${slugify(value)}/`}
           >
             {value}
           </Link>
