+++ src/components/ArmorColumnProvider.js
@@ -10,6 +10,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function ArmorColumnProvider({ children, currentBook }) {
   let bookData = ProvideBookData()
   let columns = makeColumns(
@@ -19,7 +21,11 @@ export default function ArmorColumnProvider({ children, currentBook }) {
         name: "name",
         options: {
           customBodyRender: (value, tableMeta) => (
-            <Link to={`/armor/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}>
+            <Link
+              to={`/armor/${
+                tableMeta.rowData[GENERATED_ID_COL_INDEX]
+              }/${slugify(value)}/`}
+            >
               {value}
             </Link>
           ),
@@ -42,10 +48,9 @@ export default function ArmorColumnProvider({ children, currentBook }) {
         name: "price",
         options: {
           customBodyRender: (value, tableMeta) =>
-            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${(value &&
-              value.toLocaleString &&
-              value.toLocaleString()) ||
-              value}`,
+            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${
+              (value && value.toLocaleString && value.toLocaleString()) || value
+            }`,
           ...PRICE_FILTER_OPTIONS,
         },
       },
