+++ src/components/SpeciesColumnProvider.js
@@ -8,6 +8,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function SpeciesColumnProvider({ children, currentBook }) {
   let bookData = ProvideBookData()
   let columns = makeColumns([
@@ -16,7 +18,11 @@ export default function SpeciesColumnProvider({ children, currentBook }) {
       name: "name",
       options: {
         customBodyRender: (value, tableMeta) => (
-          <Link to={`/species/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}>
+          <Link
+            to={`/species/${
+              tableMeta.rowData[GENERATED_ID_COL_INDEX]
+            }/${slugify(value)}/`}
+          >
             {value}
           </Link>
         ),
