+++ src/components/StarshipsColumnProvider.js
@@ -10,6 +10,8 @@ import {
 } from "./shared/ColumnHelper"
 import ProvideBookData from "./shared/BookDataProvider"
 
+import { slugify } from "../utils/slugify"
+
 export default function StarshipsColumnProvider({ children, currentBook }) {
   let bookData = ProvideBookData()
   let columns = makeColumns(
@@ -20,7 +22,9 @@ export default function StarshipsColumnProvider({ children, currentBook }) {
         options: {
           customBodyRender: (value, tableMeta) => (
             <Link
-              to={`/starships/${tableMeta.rowData[GENERATED_ID_COL_INDEX]}/`}
+              to={`/starships/${
+                tableMeta.rowData[GENERATED_ID_COL_INDEX]
+              }/${slugify(value)}/`}
             >
               {value}
             </Link>
@@ -84,10 +88,9 @@ export default function StarshipsColumnProvider({ children, currentBook }) {
         name: "price",
         options: {
           customBodyRender: (value, tableMeta) =>
-            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${(value &&
-              value.toLocaleString &&
-              value.toLocaleString()) ||
-              value}`,
+            `${tableMeta.rowData[RESTRICTED_COL_INDEX] ? "(R) " : ""}${
+              (value && value.toLocaleString && value.toLocaleString()) || value
+            }`,
           ...PRICE_FILTER_OPTIONS,
         },
       },
